ipipe: add cpuidle control interface
The co-kernel hook which should determine whether entering a sleep
state is accepted is renamed to ipipe_cpuidle_control(), now receiving
the cpuidle device and state information for enabling a smarter logic
on the client side.

ipipe_enter_idle_hook() that might be implemented by co-kernels
conforming to the older interface won't be called anymore, which
defaults to accepting the transition to idle
unconditionally. Disabling CONFIG_CPU_IDLE or passing idle=poll would
restore the original behavior, which is still the current
recommendation to Xenomai users anyway.