Revert "arm/ipipe: notify about unrecoverable undefined instruction only"
This reverts commit ae394d544db60bba97c6e9f502a6d62472426b5e.

This change won't work with legacy co-kernels still dealing with FPU
management symmetrically to the host kernel. Those need to be informed
before call_fpe() is fired, and may want a fast exit to take place in
order to skip it.

Latency won't increase if those co-kernels do the right thing by
fixing up the FP situation directly from the head stage instead of
propagating the exception to the host kernel.