genirq: ipipe: protect generic chip against domain preemption
As described in Documentation/ipipe.rst, irq_chip drivers need to be
specifically adapted for dealing with interrupt pipelining safely.

The basic issue to address is proper serialization between some
irq_chip handlers which may be called from out-of-band context
immediately upon receipt of an IRQ, and the rest of the driver which
may access the same data / IO registers from the regular - in-band -
context on the same CPU.

This commit converts the generic irq_chip lock to a hard spinlock,
which ensures such serialization.