timekeeping: ipipe: forward clock shift value to DSO helpers
In order to propagate the "host real-time update" event to a co-kernek
(IPIPE_KEVT_HOSTRT), we need the clock shift value of the monotonic
clock to be passed to the legacy vDSO handler, for (re)calculating the
new wall clock time which is eventually announced to the co-kernel.

Only architectures which still implement the legacy
update_vsyscall_old() interface need this change.