ipipe: add kernel event notifiers
Add the core API for enabling (regular) kernel event notifications to
a co-kernel running over the head domain. For instance, such a
co-kernel may need to know when a task is about to be resumed upon
signal receipt, or when it gets an access fault trap.

This commit adds the client-side API for enabling such notification
for class of events, but does not provide the notification points per
se, which comes later.