ipipe: timer: resume hardware operations in oneshot handler
Although we won't allow disabling the hardware when the clock event
logic switches a device to stopped mode - so that we won't affect the
timer logic running on the head stage unexpectedly -, we still have to
enable the hardware when switched (back) to oneshot mode, since it may
have been stopped prior to interposing on the device in
ipipe_timer_start().

Failing to do so would leave the hardware shut down for both regular
and Xenomai operations, with no mean to bring it up again.