ipipe: switch potentially large cpumask to static storage
When a huge number of CPUs is available (e.g. CONFIG_MAXSMP/x86), we
might overflow the stack with cpumask_t variables in
ipipe_critical_enter().

Instead of allocating cpumask_var_t dynamically for these, rely on the
fact that we cannot reenter the code accessing them by design, so
those variables may be moved to local static storage.