arm: ipipe: Fix up the omap-gpio driver to deliver interrupts properly to the pipeline
OMAP gpio driver was changed upstream to be able to support threaded IRQ handler,
this change impacted how the ipipe dealt with this driver. Since the ipipe expects
a chained handler and now it's a generic one, we need to ack the event in the
interrupt controller before calling the handler and unmask it after.

Level flow handler must be used for pipelined interrupts. Since the IRQ handler
 on the root stage may be delayed until the real-time core releases the CPU, we
need to mask the IRQ to prevent an IRQ storm when the interrupts are enabled again.

Convert wa_lock to ipipe_spinlock, since wa_lock is used in the irq_chip functions
which when called from the head stage can't use a Linux kernel service.

Add the hold/release handlers for robustness.

Signed-off-by: Greg Gallagher <greg@embeddedgreg.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>