ipipe: arm: arm_global_timer: Use 64 Bit hw counter for ipipe_tsc
When using the arm_global_timer as source for ipipe_tsc the mask
was configured with 32 Bit.
However this timer is a native 64 Bit counter. So use the complete
hardware register by configuring a 64 Bit mask on initialization.
The ipipe_tsc implementation is already 64 Bit ready and reads this
register as described in the reference manual, so no further
changes needed.

Signed-off-by: Gunter Grau <gunter.grau@philips.com>
Signed-off-by: Greg Gallagher <greg@embeddedgreg.com>